+++ hazelcast-raft/src/test/java/com/hazelcast/raft/impl/service/RaftDataService.java
@@ -10,16 +10,16 @@ import java.util.Map.Entry;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class RaftDataService implements SnapshotAwareService<Map<Integer, Object>> {
+public class RaftDataService implements SnapshotAwareService<Map<Long, Object>> {
 
     public static final String SERVICE_NAME = "RaftTestService";
 
-    private final Map<Integer, Object> values = new ConcurrentHashMap<Integer, Object>();
+    private final Map<Long, Object> values = new ConcurrentHashMap<Long, Object>();
 
     public RaftDataService() {
     }
 
-    Object apply(int commitIndex, Object value) {
+    Object apply(long commitIndex, Object value) {
         assert !values.containsKey(commitIndex) :
                 "Cannot apply " + value + "since commitIndex: " + commitIndex + " already contains: " + values.get(commitIndex);
 
@@ -27,7 +27,7 @@ public class RaftDataService implements SnapshotAwareService<Map<Integer, Object
         return value;
     }
 
-    public Object get(int commitIndex) {
+    public Object get(long commitIndex) {
         return values.get(commitIndex);
     }
 
@@ -40,9 +40,9 @@ public class RaftDataService implements SnapshotAwareService<Map<Integer, Object
     }
 
     @Override
-    public Map<Integer, Object> takeSnapshot(RaftGroupId raftGroupId, int commitIndex) {
-        Map<Integer, Object> snapshot = new HashMap<Integer, Object>();
-        for (Entry<Integer, Object> e : values.entrySet()) {
+    public Map<Long, Object> takeSnapshot(RaftGroupId raftGroupId, long commitIndex) {
+        Map<Long, Object> snapshot = new HashMap<Long, Object>();
+        for (Entry<Long, Object> e : values.entrySet()) {
             assert e.getKey() <= commitIndex : "Key: " + e.getKey() + ", commit-index: " + commitIndex;
             snapshot.put(e.getKey(), e.getValue());
         }
@@ -51,7 +51,7 @@ public class RaftDataService implements SnapshotAwareService<Map<Integer, Object
     }
 
     @Override
-    public void restoreSnapshot(RaftGroupId raftGroupId, int commitIndex, Map<Integer, Object> snapshot) {
+    public void restoreSnapshot(RaftGroupId raftGroupId, long commitIndex, Map<Long, Object> snapshot) {
         values.clear();
         values.putAll(snapshot);
     }
